requirements-tests.txt
-r requirements.txt
-e .

muffin-jinja2
pytest
